Memory Stuff, idk

+++ source/Engine/Renderer/Renderer.h
@@ -16,7 +16,7 @@ namespace antares {
 		void BeginFrame();
 		void EndFrame();
 
-		void SetColor(int r, int g, int b, int a);
+		void SetColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
 		void DrawPoint(int x, int y);
 		void DrawPoint(float x, float y) { DrawPoint((int)x, (int)y); }
 		void DrawLine(int x1, int y1, int x2, int y2);
